test(openclaw): replace opportunity-pickup spec with opportunity-batch spec
